Moving method containing logic out of the cHomeController class
