+++ src/Server/Controllers/HomeController.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Threading.Tasks;
-using static EctBlazorApp.Server.CommonMethods.GraphMethods;
+using static EctBlazorApp.Server.Behaviour.UserBehaviour;
 
 namespace EctBlazorApp.Server.Controllers
 {
@@ -28,31 +28,11 @@ namespace EctBlazorApp.Server.Controllers
         {
             using var client = new HttpClient();
             client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("bearer", graphToken);
-            EctUser userForParms = await GetExistingEctUserOrNew(userId, client, _dbContext);
+            EctUser userForParms = await GetExistingEctUserOrNewWrapper(userId, client, _dbContext);
             // update calendarEvents
             // update receivedMail
             // update sentMail
             return BadRequest();
         }
-
-        private static async Task<EctUser> GetExistingEctUserOrNew(string userId, HttpClient client, EctDbContext dbContext)
-        {
-            try
-            {
-                EctUser userForUserIdParm = dbContext.Users.First(user => user.Email.Equals(userId));
-                return userForUserIdParm;
-            }
-            catch (Exception)
-            {
-                EctUser addUserResult = await dbContext.AddUser(userId, client);
-                return addUserResult;
-            }
-        }
-
-        public static Task<EctUser> GetExistingEctUserOrNewWrapper(string userId, HttpClient client, EctDbContext dbContext)
-        {
-            return GetExistingEctUserOrNew(userId, client, dbContext);
-        }
-
     }
 }
